requirement.txt
amqp==2.5.2
billiard==3.6.3.0
celery==4.4.1
certifi==2019.11.28
chardet==3.0.4
Django==2.2
djangorestframework==3.11.0
entrypoints==0.3
environ==1.0
environs==7.3.0
flake8==3.7.9
idna==2.9
importlib-metadata==1.5.0
kombu==4.6.8
marshmallow==3.5.1
mccabe==0.6.1
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
python-dotenv==0.12.0
pytz==2019.3
redis==3.4.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
vine==1.3.0
zipp==3.1.0
